refactor into python module
